function/requirements.txt
appdirs==1.4.4
artifacts-keyring==0.3.1
azure-common==1.1.27
azure-core==1.20.1
azure-functions==1.8.0
azure-servicebus==7.4.0
black==21.7b0
bleach==3.3.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
distlib==0.3.2
distro==1.5.0
docutils==0.17.1
dotnetcore2==2.1.11
filelock==3.0.12
Flask==1.1.2
google-api-core==1.31.2
google-auth==1.35.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
idna==3.2
importlib-metadata==4.6.1
isodate==0.6.0
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
keyring==23.0.1
MarkupSafe==2.0.1
msrest==0.6.21
mypy-extensions==0.4.3
oauthlib==3.1.1
opencensus==0.7.13
opencensus-context==0.1.2
opencensus-ext-azure==1.0.4
opencensus-ext-flask==0.7.3
opencensus-ext-logging==0.1.0
packaging==21.0
pathspec==0.9.0
pipenv==2021.5.29
pkginfo==1.7.1
protobuf==3.17.3
psutil==5.8.0
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
pyparsing==2.4.7
pytz==2021.1
readme-renderer==29.0
redis==3.5.3
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.27
tomli==1.2.1
tqdm==4.61.2
twine==3.4.1
uamqp==1.4.3
urllib3==1.26.6
virtualenv==20.4.7
virtualenv-clone==0.5.4
webencodings==0.5.1
Werkzeug==2.0.1
zipp==3.5.0
